requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
itsdangerous==2.1.2
requests==2.31.0